usb_audio: extract UsbRawSinkConfig to raw_config (no GStreamer deps)
Prep step before deleting the V1 GStreamer raw sink: the struct itself
is shared with native_transport (V2), but its previous home pulled in
gstreamer/glib via raw_sink.rs. Move it into a stand-alone raw_config
module so V2 can keep using it after raw_sink.rs is removed.
