+++ src_rust/rust_audio_core/src/usb_audio/mod.rs
@@ -3,23 +3,20 @@
 //! # Architecture
 //!
 //! ```text
-//! GStreamer appsink (pull mode)
-//!        │  PCM frames (S32LE / S24LE / S16LE)
+//! Symphonia decoder (PCM slabs)
+//!        │  S16LE / S24_3LE / S32LE / F32LE / F64LE
 //!   FrameQueue  (SPSC lock-free ring, ~200 ms)
 //!        │
 //!  UsbAudioThread  (dedicated RT thread)
 //!   ├── IsoTransferRing   (libusb async iso, N=16 transfers)
 //!   ├── FeedbackReader    (UAC 2.0 async sample-rate feedback)
 //!   ├── RateAdapter       (dynamic per-frame sample count)
-//!   └── HwClockFeed       (frame counter → AlsaHwClock)
+//!   └── HwClockFeed       (frame counter → master clock)
 //!        │
 //!   libusb-1.0  (async, event loop in dedicated thread)
 //!        │  USB Isochronous OUT
 //!   USB DAC
 //! ```
-//!
-//! Phase 1 (this file): descriptor parsing + device enumeration.
-//! Later phases add the transfer ring, feedback reader, and Engine integration.
 
 pub mod control;
 pub mod borrowed_queue;
@@ -29,6 +26,7 @@ pub mod dop;
 pub mod feedback;
 pub mod owned_buffer_queue;
 pub mod queue;
+pub mod raw_config;
 pub mod raw_sink;
 pub mod sink;
 pub mod source;
@@ -40,7 +38,8 @@ pub use device::{enumerate_usb_audio_devices, OpenUsbDevice, UacAltProfile, UsbA
 pub use dop::DopEncoder;
 pub use owned_buffer_queue::OwnedBufferQueue;
 pub use queue::FrameQueue;
-pub use raw_sink::{UsbRawSink, UsbRawSinkConfig};
+pub use raw_config::UsbRawSinkConfig;
+pub use raw_sink::UsbRawSink;
 pub use sink::{QueueMode, UsbAudioSink};
 pub use source::TransferSource;
 pub use transfer::{IsoTransferRing, RingState};
